remove globals from window_background
