+++ window_background/background-util.js
@@ -1,13 +1,9 @@
-/*
-global
-  debugLog
-  firstPass
-*/
 // Utility functions that belong only to background
 const { ipcRenderer: ipc } = require("electron");
 const _ = require("lodash");
 const parse = require("date-fns/parse");
 const isValid = require("date-fns/isValid");
+const globals = require("./globals");
 
 const {
   IPC_BACKGROUND,
@@ -149,7 +145,7 @@ const overlayWhitelist = [
 
 // convenience fn to update player data singletons in all processes
 // (update is destructive, be sure to use spread syntax if necessary)
-function setData(data, refresh = debugLog || !firstPass) {
+function setData(data, refresh = globals.debugLog || !global.firstPass) {
   const cleanData = _.omit(data, dataBlacklist);
   playerData.handleSetData(null, JSON.stringify(cleanData));
   ipc_send("set_player_data", JSON.stringify(cleanData), IPC_MAIN);
