+++ window_background/labels.js
@@ -1,42 +1,6 @@
 /*
   global
-    logTime
-    actionLog
-    skipMatch
-    currentMatch
-    getNameBySeat
-    logLanguage
-    gameNumberCompleted
-    initialLibraryInstanceIds
-    instanceToCardIdMap
-    idChanges
-    matchGameStats
-    saveMatch
-    greToClientInterpreter
-    decodePayload
-    addCustomDeck
-    saveCourse
-    select_deck
-    sha1
-    store
-    saveEconomyTransaction
-    matchBeginTime
-    clearDraftData
-    processMatch
-    getDraftData
-    endDraft
-    setDraftData
-    startDraft
-    clear_deck
-    duringMatch
-    playerWin
-    draws
-    oppWin
-    matchCompletedOnGameNumber
-    oppId
-    playerData
-    firstPass
-    debugLog
+    clear_dec
 */
 const _ = require("lodash");
 const differenceInDays = require("date-fns/differenceInDays");
@@ -46,6 +10,13 @@ const db = require("../shared/database");
 const CardsList = require("../shared/cards-list");
 const { get_deck_colors, objectClone, replaceAll } = require("../shared/util");
 
+const greToClientInterpreter = require("./gre-to-client-interpreter");
+const playerData = require("../shared/player-data.js");
+const sha1 = require("js-sha1");
+const httpApi = require("./http-api");
+const getNameBySeat = require("./getNameBySeat");
+const Deck = require("../shared/deck");
+
 const {
   httpSetMythicRank,
   httpSubmitCourse,
@@ -54,10 +25,258 @@ const {
 const {
   ipc_send,
   normaliseFields,
-  parseWotcTime,
   parseWotcTimeFallback,
   setData
 } = require("./background-util");
+const { IPC_OVERLAY } = require("../shared/constants.js");
+
+const actionLog = require("./actionLog");
+const addCustomDeck = require("./addCustomDeck");
+
+const globals = require("./globals");
+
+const { createDraft, createMatch, completeMatch } = require("./data");
+
+var logLanguage = "English";
+
+let skipMatch = false;
+
+function clear_deck() {
+  var deck = { mainDeck: [], sideboard: [], name: "" };
+  ipc_send("set_deck", deck, IPC_OVERLAY);
+}
+
+function clearDraftData(draftId) {
+  if (playerData.draftExists(draftId)) {
+    if (playerData.draft_index.includes(draftId)) {
+      const draft_index = [...playerData.draft_index];
+      draft_index.splice(draft_index.indexOf(draftId), 1);
+      setData({ draft_index }, false);
+      if (globals.debugLog || !globals.firstPass)
+        globals.store.set("draft_index", draft_index);
+    }
+    setData({ [draftId]: null });
+    // Note: we must always run delete, regardless of firstpass
+    globals.store.delete(draftId);
+  }
+}
+
+function decodePayload(json) {
+  const messages = require("./messages_pb");
+
+  const msgType = json.clientToMatchServiceMessageType.split("_")[1],
+    binaryMsg = new Buffer.from(json.payload, "base64");
+
+  try {
+    let msgDeserialiser;
+    if (
+      msgType === "ClientToGREMessage" ||
+      msgType === "ClientToGREUIMessage"
+    ) {
+      msgDeserialiser = messages.ClientToGREMessage;
+    } else if (msgType === "ClientToMatchDoorConnectRequest") {
+      msgDeserialiser = messages.ClientToMatchDoorConnectRequest;
+    } else if (msgType === "AuthenticateRequest") {
+      msgDeserialiser = messages.AuthenticateRequest;
+    } else if (msgType === "CreateMatchGameRoomRequest") {
+      msgDeserialiser = messages.CreateMatchGameRoomRequest;
+    } else if (msgType === "EchoRequest") {
+      msgDeserialiser = messages.EchoRequest;
+    } else {
+      console.warn(`${msgType} - unknown message type`);
+      return;
+    }
+    const msg = msgDeserialiser.deserializeBinary(binaryMsg);
+    //console.log(json.msgType);
+    //console.log(msg.toObject());
+    return msg.toObject();
+  } catch (e) {
+    console.log(e.message);
+  }
+
+  return;
+}
+
+function startDraft() {
+  if (globals.debugLog || !globals.firstPass) {
+    if (playerData.settings.close_on_match) {
+      ipc_send("renderer_hide", 1);
+    }
+    ipc_send("set_arena_state", ARENA_MODE_DRAFT);
+  }
+  globals.duringDraft = true;
+}
+
+function getDraftData(id, entry) {
+  var data = playerData.draft(id) || createDraft(id, entry);
+
+  if (!data.date && entry.timestamp) {
+    // the first event we see we set the date.
+    data.timestamp = entry.timestamp;
+    data.date = parseWotcTimeFallback(entry.timestamp);
+  }
+
+  return data;
+}
+
+//
+function setDraftData(data) {
+  if (!data || !data.id) {
+    console.log("Couldnt save undefined draft:", data);
+    return;
+  }
+  const { id } = data;
+
+  // console.log("Set draft data:", data);
+  if (!playerData.draft_index.includes(id)) {
+    const draft_index = [...playerData.draft_index, id];
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("draft_index", draft_index);
+    setData({ draft_index }, false);
+  }
+
+  if (globals.debugLog || !globals.firstPass) globals.store.set(id, data);
+  setData({
+    [id]: data,
+    cards: playerData.cards,
+    cardsNew: playerData.cardsNew
+  });
+  ipc_send("set_draft_cards", data, IPC_OVERLAY);
+}
+
+//
+function endDraft(data) {
+  globals.duringDraft = false;
+  if (globals.debugLog || !globals.firstPass)
+    ipc_send("set_arena_state", ARENA_MODE_IDLE);
+  if (!data) return;
+  httpApi.httpSetDraft(data);
+  ipc_send("popup", { text: "Draft saved!", time: 3000 });
+}
+
+// Create a match from data, set globals and trigger ipc
+function processMatch(json, matchBeginTime) {
+  actionLog(-99, new Date(), "");
+
+  if (globals.debugLog || !globals.firstPass) {
+    ipc_send("set_arena_state", ARENA_MODE_MATCH);
+  }
+
+  var match = createMatch(json, matchBeginTime);
+
+  // set global values
+
+  globals.currentMatch = match;
+  globals.matchGameStats = [];
+  globals.matchCompletedOnGameNumber = 0;
+  globals.gameNumberCompleted = 0;
+  globals.initialLibraryInstanceIds = [];
+  globals.idChanges = {};
+  globals.instanceToCardIdMap = {};
+
+  ipc_send("ipc_log", "vs " + match.opponent.name);
+  ipc_send("set_timer", match.beginTime, IPC_OVERLAY);
+
+  if (match.eventId == "DirectGame" && globals.currentDeck) {
+    let str = globals.currentDeck.getSave();
+    httpApi.httpTournamentCheck(str, match.opponent.name, true);
+  }
+
+  ipc_send("set_priority_timer", match.priorityTimers, IPC_OVERLAY);
+
+  return match;
+}
+
+//
+function saveCourse(json) {
+  const id = json._id;
+  delete json._id;
+  json.id = id;
+  const eventData = {
+    date: new Date(),
+    // preserve custom fields if possible
+    ...(playerData.event(id) || {}),
+    ...json
+  };
+
+  if (!playerData.courses_index.includes(id)) {
+    const courses_index = [...playerData.courses_index, id];
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("courses_index", courses_index);
+    setData({ courses_index }, false);
+  }
+
+  if (globals.debugLog || !globals.firstPass) globals.store.set(id, eventData);
+  setData({ [id]: eventData });
+}
+
+function saveEconomyTransaction(transaction) {
+  const id = transaction.id;
+  const txnData = {
+    // preserve custom fields if possible
+    ...(playerData.transaction(id) || {}),
+    ...transaction
+  };
+
+  if (!playerData.economy_index.includes(id)) {
+    const economy_index = [...playerData.economy_index, id];
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("economy_index", economy_index);
+    setData({ economy_index }, false);
+  }
+
+  if (globals.debugLog || !globals.firstPass) globals.store.set(id, txnData);
+  setData({ [id]: txnData });
+  httpApi.httpSetEconomy(txnData);
+}
+
+function saveMatch(id, matchEndTime) {
+  //console.log(globals.currentMatch.matchId, id);
+  if (
+    !globals.currentMatch ||
+    !globals.currentMatch.matchTime ||
+    globals.currentMatch.matchId !== id
+  ) {
+    return;
+  }
+  const existingMatch = playerData.match(id) || {};
+  const match = completeMatch(
+    existingMatch,
+    globals.currentMatch,
+    matchEndTime
+  );
+  if (!match) {
+    return;
+  }
+
+  // console.log("Save match:", match);
+  if (!playerData.matches_index.includes(id)) {
+    const matches_index = [...playerData.matches_index, id];
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("matches_index", matches_index);
+    setData({ matches_index }, false);
+  }
+
+  if (globals.debugLog || !globals.firstPass) globals.store.set(id, match);
+  setData({ [id]: match });
+  if (globals.matchCompletedOnGameNumber === globals.gameNumberCompleted) {
+    httpApi.httpSetMatch(match);
+  }
+  ipc_send("set_timer", 0, IPC_OVERLAY);
+  ipc_send("popup", { text: "Match saved!", time: 3000 });
+}
+
+function select_deck(arg) {
+  if (arg.CourseDeck) {
+    globals.currentDeck = new Deck(arg.CourseDeck);
+  } else {
+    globals.currentDeck = new Deck(arg);
+  }
+  // console.log("Select deck: ", globals.currentDeck, arg);
+  globals.originalDeck = globals.currentDeck.clone();
+  ipc_send("set_deck", globals.currentDeck.getSave(), IPC_OVERLAY);
+}
+
 //
 function convert_deck_from_v3(deck) {
   return JSON.parse(JSON.stringify(deck), (key, value) => {
@@ -76,7 +295,7 @@ function convert_deck_from_v3(deck) {
 
 function onLabelOutLogInfo(entry, json) {
   if (!json) return;
-  logTime = parseWotcTimeFallback(entry.timestamp);
+  globals.logTime = parseWotcTimeFallback(entry.timestamp);
 
   if (json.params.messageName == "Client.UserDeviceSpecs") {
     let payload = {
@@ -88,7 +307,7 @@ function onLabelOutLogInfo(entry, json) {
   }
   if (json.params.messageName == "DuelScene.GameStart") {
     let gameNumber = json.params.payloadObject.gameNumber;
-    actionLog(-2, logTime, `Game ${gameNumber} Start`);
+    actionLog(-2, globals.logTime, `Game ${gameNumber} Start`);
   }
 
   if (json.params.messageName == "Client.Connected") {
@@ -96,43 +315,43 @@ function onLabelOutLogInfo(entry, json) {
   }
   if (skipMatch) return;
   if (json.params.messageName == "DuelScene.GameStop") {
-    currentMatch.opponent.cards = currentMatch.oppCardsUsed;
+    globals.currentMatch.opponent.cards = globals.currentMatch.oppCardsUsed;
 
     var payload = json.params.payloadObject;
 
     let loserName = getNameBySeat(payload.winningTeamId == 1 ? 2 : 1);
     if (payload.winningReason == "ResultReason_Concede") {
-      actionLog(-1, logTime, `${loserName} Conceded`);
+      actionLog(-1, globals.logTime, `${loserName} Conceded`);
     }
     if (payload.winningReason == "ResultReason_Timeout") {
-      actionLog(-1, logTime, `${loserName} Timed out`);
+      actionLog(-1, globals.logTime, `${loserName} Timed out`);
     }
 
     let playerName = getNameBySeat(payload.winningTeamId);
-    actionLog(-1, logTime, `${playerName} Wins!`);
+    actionLog(-1, globals.logTime, `${playerName} Wins!`);
 
     var mid = payload.matchId + "-" + playerData.arenaId;
     var time = payload.secondsCount;
-    if (mid == currentMatch.matchId) {
-      gameNumberCompleted = payload.gameNumber;
+    if (mid == globals.currentMatch.matchId) {
+      globals.gameNumberCompleted = payload.gameNumber;
 
       let game = {};
       game.time = time;
       game.winner = payload.winningTeamId;
       // Just a shortcut for future aggregations
-      game.win = payload.winningTeamId == currentMatch.player.seat;
+      game.win = payload.winningTeamId == globals.currentMatch.player.seat;
 
       game.shuffledOrder = [];
-      for (let i = 0; i < initialLibraryInstanceIds.length; i++) {
-        let instance = initialLibraryInstanceIds[i];
+      for (let i = 0; i < globals.initialLibraryInstanceIds.length; i++) {
+        let instance = globals.initialLibraryInstanceIds[i];
         while (
-          (!instanceToCardIdMap[instance] ||
-            !db.card(instanceToCardIdMap[instance])) &&
-          idChanges[instance]
+          (!globals.instanceToCardIdMap[instance] ||
+            !db.card(globals.instanceToCardIdMap[instance])) &&
+          globals.idChanges[instance]
         ) {
-          instance = idChanges[instance];
+          instance = globals.idChanges[instance];
         }
-        let cardId = instanceToCardIdMap[instance];
+        let cardId = globals.instanceToCardIdMap[instance];
         if (db.card(cardId)) {
           game.shuffledOrder.push(cardId);
         } else {
@@ -146,15 +365,17 @@ function onLabelOutLogInfo(entry, json) {
         game.shuffledOrder.slice(0, 7 - game.handsDrawn.length)
       );
 
-      if (gameNumberCompleted > 1) {
+      if (globals.gameNumberCompleted > 1) {
         let deckDiff = {};
-        currentMatch.player.deck.mainboard.get().forEach(card => {
+        globals.currentMatch.player.deck.mainboard.get().forEach(card => {
           deckDiff[card.id] = card.quantity;
         });
-        currentMatch.player.originalDeck.mainboard.get().forEach(card => {
-          deckDiff[card.id] = (deckDiff[card.id] || 0) - card.quantity;
-        });
-        matchGameStats.forEach((stats, i) => {
+        globals.currentMatch.player.originalDeck.mainboard
+          .get()
+          .forEach(card => {
+            deckDiff[card.id] = (deckDiff[card.id] || 0) - card.quantity;
+          });
+        globals.matchGameStats.forEach((stats, i) => {
           if (i !== 0) {
             let prevChanges = stats.sideboardChanges;
             prevChanges.added.forEach(
@@ -181,7 +402,7 @@ function onLabelOutLogInfo(entry, json) {
         });
 
         game.sideboardChanges = sideboardChanges;
-        game.deck = objectClone(currentMatch.player.deck.getSave());
+        game.deck = objectClone(globals.currentMatch.player.deck.getSave());
       }
 
       game.handLands = game.handsDrawn.map(
@@ -192,7 +413,7 @@ function onLabelOutLogInfo(entry, json) {
       let landsInDeck = 0;
       let multiCardPositions = { "2": {}, "3": {}, "4": {} };
       let cardCounts = {};
-      currentMatch.player.deck.mainboard.get().forEach(card => {
+      globals.currentMatch.player.deck.mainboard.get().forEach(card => {
         cardCounts[card.id] = card.quantity;
         deckSize += card.quantity;
         if (card.quantity >= 2 && card.quantity <= 4) {
@@ -222,8 +443,8 @@ function onLabelOutLogInfo(entry, json) {
         }
       });
 
-      game.cardsCast = _.cloneDeep(currentMatch.cardsCast);
-      currentMatch.cardsCast = [];
+      game.cardsCast = _.cloneDeep(globals.currentMatch.cardsCast);
+      globals.currentMatch.cardsCast = [];
       game.deckSize = deckSize;
       game.landsInDeck = landsInDeck;
       game.multiCardPositions = multiCardPositions;
@@ -231,11 +452,11 @@ function onLabelOutLogInfo(entry, json) {
       game.landsInLibrary = landsInLibrary;
       game.libraryLands = libraryLands;
 
-      currentMatch.matchTime = matchGameStats.reduce(
+      globals.currentMatch.matchTime = globals.matchGameStats.reduce(
         (acc, cur) => acc + cur.time,
         0
       );
-      matchGameStats[gameNumberCompleted - 1] = game;
+      globals.matchGameStats[globals.gameNumberCompleted - 1] = game;
 
       saveMatch(mid);
     }
@@ -243,8 +464,9 @@ function onLabelOutLogInfo(entry, json) {
   if (json.params.messageName === "Client.SceneChange") {
     const { toSceneName } = json.params.payloadObject;
     if (toSceneName === "Home") {
-      if (debugLog || !firstPass) ipc_send("set_arena_state", ARENA_MODE_IDLE);
-      duringMatch = false;
+      if (globals.debugLog || !globals.firstPass)
+        ipc_send("set_arena_state", ARENA_MODE_IDLE);
+      globals.duringMatch = false;
       endDraft();
     }
   }
@@ -253,16 +475,16 @@ function onLabelOutLogInfo(entry, json) {
 function onLabelGreToClient(entry, json) {
   if (!json) return;
   if (skipMatch) return;
-  logTime = parseWotcTimeFallback(entry.timestamp);
+  globals.logTime = parseWotcTimeFallback(entry.timestamp);
 
   json = json.greToClientEvent.greToClientMessages;
   json.forEach(function(msg) {
     let msgId = msg.msgId;
-    greToClientInterpreter.GREMessage(msg, logTime);
+    greToClientInterpreter.GREMessage(msg, globals.logTime);
     /*
-    currentMatch.GREtoClient[msgId] = msg;
-    currentMatch.latestMessage = msgId;
-    greToClientInterpreter.GREMessageByID(msgId, logTime);
+    globals.currentMatch.GREtoClient[msgId] = msg;
+    globals.currentMatch.latestMessage = msgId;
+    greToClientInterpreter.GREMessageByID(msgId, globals.logTime);
     */
   });
 }
@@ -287,13 +509,13 @@ function onLabelClientToMatchServiceMessageTypeClientToGREMessage(entry, json) {
 
     let tempMain = new CardsList(deckResp.deckcards);
     let tempSide = new CardsList(deckResp.sideboardcards);
-    let newDeck = currentMatch.player.deck.clone();
+    let newDeck = globals.currentMatch.player.deck.clone();
     newDeck.mainboard = tempMain;
     newDeck.sideboard = tempSide;
     newDeck.getColors();
 
-    currentMatch.player.deck = newDeck;
-    console.log("> ", currentMatch.player.deck);
+    globals.currentMatch.player.deck = newDeck;
+    console.log("> ", globals.currentMatch.player.deck);
   }
 }
 
@@ -334,8 +556,8 @@ function onLabelInEventGetCombinedRankInfo(entry, json) {
   }
 
   setData({ rank });
-  if (debugLog || !firstPass) {
-    store.set("rank", rank);
+  if (globals.debugLog || !globals.firstPass) {
+    globals.store.set("rank", rank);
   }
 }
 
@@ -360,8 +582,8 @@ function onLabelRankUpdated(entry, json) {
   rank[updateType].seasonOrdinal = json.seasonOrdinal;
 
   setData({ rank });
-  if (debugLog || !firstPass) {
-    store.set("rank", rank);
+  if (globals.debugLog || !globals.firstPass) {
+    globals.store.set("rank", rank);
   }
 }
 
@@ -384,8 +606,8 @@ function onLabelMythicRatingUpdated(entry, json) {
   rank.constructed.leaderboardPlace = json.newMythicLeaderboardPlacement;
 
   setData({ rank });
-  if (debugLog || !firstPass) {
-    store.set("rank", rank);
+  if (globals.debugLog || !globals.firstPass) {
+    globals.store.set("rank", rank);
   }
 }
 
@@ -397,12 +619,14 @@ function onLabelInDeckGetDeckLists(entry, json) {
   json.forEach(deck => {
     const deckData = { ...(playerData.deck(deck.id) || {}), ...deck };
     decks[deck.id] = deckData;
-    if (debugLog || !firstPass) store.set("decks." + deck.id, deckData);
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("decks." + deck.id, deckData);
     static_decks.push(deck.id);
   });
 
   setData({ decks, static_decks });
-  if (debugLog || !firstPass) store.set("static_decks", static_decks);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("static_decks", static_decks);
 }
 
 function onLabelInDeckGetDeckListsV3(entry, json) {
@@ -430,7 +654,8 @@ function onLabelInEventGetPlayerCourses(entry, json) {
   });
 
   setData({ static_events });
-  if (debugLog || !firstPass) store.set("static_events", static_events);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("static_events", static_events);
 }
 
 function onLabelInEventGetPlayerCoursesV2(entry, json) {
@@ -549,22 +774,23 @@ function onLabelInDeckUpdateDeck(entry, json) {
     (deltaDeck.changesMain.length || deltaDeck.changesSide.length);
 
   if (foundNewDeckChange) {
-    if (debugLog || !firstPass)
-      store.set("deck_changes." + changeId, deltaDeck);
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("deck_changes." + changeId, deltaDeck);
     const deck_changes = { ...playerData.deck_changes, [changeId]: deltaDeck };
     const deck_changes_index = [...playerData.deck_changes_index];
     if (!deck_changes_index.includes(changeId)) {
       deck_changes_index.push(changeId);
     }
-    if (debugLog || !firstPass)
-      store.set("deck_changes_index", deck_changes_index);
+    if (globals.debugLog || !globals.firstPass)
+      globals.store.set("deck_changes_index", deck_changes_index);
 
     setData({ deck_changes, deck_changes_index });
   }
 
   const deckData = { ..._deck, ...json };
   const decks = { ...playerData.decks, [json.id]: deckData };
-  if (debugLog || !firstPass) store.set("decks." + json.id, deckData);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("decks." + json.id, deckData);
   setData({ decks });
 }
 
@@ -616,7 +842,7 @@ function onLabelInventoryUpdated(entry, transaction) {
 
 function onLabelInPlayerInventoryGetPlayerInventory(entry, json) {
   if (!json) return;
-  logTime = parseWotcTimeFallback(entry.timestamp);
+  globals.logTime = parseWotcTimeFallback(entry.timestamp);
   const economy = {
     ...playerData.economy,
     gold: json.gold,
@@ -630,7 +856,8 @@ function onLabelInPlayerInventoryGetPlayerInventory(entry, json) {
     boosters: json.boosters
   };
   setData({ economy });
-  if (debugLog || !firstPass) store.set("economy", economy);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("economy", economy);
 }
 
 function onLabelInPlayerInventoryGetPlayerCardsV3(entry, json) {
@@ -656,7 +883,7 @@ function onLabelInPlayerInventoryGetPlayerCardsV3(entry, json) {
     cards: json
   };
 
-  if (debugLog || !firstPass) store.set("cards", cards);
+  if (globals.debugLog || !globals.firstPass) globals.store.set("cards", cards);
 
   const cardsNew = {};
   Object.keys(json).forEach(function(key) {
@@ -674,7 +901,7 @@ function onLabelInPlayerInventoryGetPlayerCardsV3(entry, json) {
 //
 function onLabelInProgressionGetPlayerProgress(entry, json) {
   if (!json || !json.activeBattlePass) return;
-  logTime = parseWotcTimeFallback(entry.timestamp);
+  globals.logTime = parseWotcTimeFallback(entry.timestamp);
   const activeTrack = json.activeBattlePass;
   const economy = {
     ...playerData.economy,
@@ -685,7 +912,8 @@ function onLabelInProgressionGetPlayerProgress(entry, json) {
     currentOrbCount: activeTrack.currentOrbCount
   };
   setData({ economy });
-  if (debugLog || !firstPass) store.set("economy", economy);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("economy", economy);
 }
 
 //
@@ -740,7 +968,8 @@ function onLabelTrackProgressUpdated(entry, json) {
   });
   // console.log(economy);
   setData({ economy });
-  if (debugLog || !firstPass) store.set("economy", economy);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("economy", economy);
 }
 
 //
@@ -781,7 +1010,8 @@ function onLabelTrackRewardTierUpdated(entry, json) {
 
   // console.log(economy);
   setData({ economy });
-  if (debugLog || !firstPass) store.set("economy", economy);
+  if (globals.debugLog || !globals.firstPass)
+    globals.store.set("economy", economy);
 }
 
 function onLabelInEventDeckSubmit(entry, json) {
@@ -926,7 +1156,7 @@ function onLabelMatchGameRoomStateChangedEvent(entry, json) {
 
   if (json.gameRoomConfig) {
     eventId = json.gameRoomConfig.eventId;
-    duringMatch = true;
+    globals.duringMatch = true;
   }
 
   if (eventId == "NPE") return;
@@ -934,7 +1164,7 @@ function onLabelMatchGameRoomStateChangedEvent(entry, json) {
   if (json.stateType == "MatchGameRoomStateType_Playing") {
     // If current match does nt exist (create match was not recieved , maybe a reconnection)
     // Only problem is recieving the decklist
-    if (!currentMatch) {
+    if (!globals.currentMatch) {
       let oName = "";
       json.gameRoomConfig.reservedPlayers.forEach(player => {
         if (!player.userId === playerData.arenaId) {
@@ -954,27 +1184,31 @@ function onLabelMatchGameRoomStateChangedEvent(entry, json) {
     }
     json.gameRoomConfig.reservedPlayers.forEach(player => {
       if (player.userId == playerData.arenaId) {
-        currentMatch.player.seat = player.systemSeatId;
+        globals.currentMatch.player.seat = player.systemSeatId;
       } else {
-        currentMatch.opponent.name = player.playerName;
-        currentMatch.opponent.id = player.userId;
-        currentMatch.opponent.seat = player.systemSeatId;
+        globals.currentMatch.opponent.name = player.playerName;
+        globals.currentMatch.opponent.id = player.userId;
+        globals.currentMatch.opponent.seat = player.systemSeatId;
       }
     });
   }
   if (json.stateType == "MatchGameRoomStateType_MatchCompleted") {
-    currentMatch.results = objectClone(json.finalMatchResult.resultList);
+    globals.currentMatch.results = objectClone(
+      json.finalMatchResult.resultList
+    );
 
     json.finalMatchResult.resultList.forEach(function(res) {
       if (res.scope == "MatchScope_Match") {
         skipMatch = false;
-        duringMatch = false;
+        globals.duringMatch = false;
       }
     });
 
     clear_deck();
-    if (debugLog || !firstPass) ipc_send("set_arena_state", ARENA_MODE_IDLE);
-    matchCompletedOnGameNumber = json.finalMatchResult.resultList.length - 1;
+    if (globals.debugLog || !globals.firstPass)
+      ipc_send("set_arena_state", ARENA_MODE_IDLE);
+    globals.matchCompletedOnGameNumber =
+      json.finalMatchResult.resultList.length - 1;
 
     var matchEndTime = parseWotcTimeFallback(entry.timestamp);
     saveMatch(
@@ -986,10 +1220,9 @@ function onLabelMatchGameRoomStateChangedEvent(entry, json) {
   if (json.players) {
     json.players.forEach(function(player) {
       if (player.userId == playerData.arenaId) {
-        currentMatch.player.seat = player.systemSeatId;
+        globals.currentMatch.player.seat = player.systemSeatId;
       } else {
-        oppId = player.userId;
-        currentMatch.opponent.seat = player.systemSeatId;
+        globals.currentMatch.opponent.seat = player.systemSeatId;
       }
     });
   }
